[FEAT] 경쟁 API 개발 - 랜덤 도전장 보내기 - 친구 도전장 보내기 - 도전장 수락 - 도전장 정보 조회 - 경쟁 인증하기(이미지) - 종료 경쟁 목록 조회 (구현 중)

+++ src/main/java/com/ssafy/sgdc/domain/entity/Badge.java
@@ -17,7 +17,7 @@ public class Badge {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // auto increment
     @Column(name = "badge_id")
-    private long badgeId;
+    private int badgeId;
 
     @Column(name = "badge_name", length = 20)
     private String badgeName;
