+++ src/main/java/com/ssafy/sgdc/domain/entity/Solo.java
@@ -1,7 +1,7 @@
 package com.ssafy.sgdc.domain.entity;
 
-import com.ssafy.sgdc.domain.entity.enums.SoloResult;
-import com.ssafy.sgdc.domain.entity.enums.SoloStatus;
+import com.ssafy.sgdc.enums.SoloResult;
+import com.ssafy.sgdc.enums.SoloStatus;
 import com.ssafy.sgdc.user.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -22,7 +22,7 @@ public class Solo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // auto increment
     @Column(name = "solo_id")
-    private long soloId;
+    private int soloId;
 
     @Column(name = "create_at")
     private LocalDateTime createAt;
